Avoid an unnecessary copy of the inserted value.

This improves the `finalize_bench.rs` by 40% - 50%, reducing the time to finalize a single `PartialEncodedChunk` from 130ms to 70ms and for `ShardChunk` it goes from 10ms to 6ms.
(Number before & after are on AMD Ryzen 9 5950X and 3200 MT/s DDR4)